001 added dynamic flavorName + code cleanup
