+++ src/main/kotlin/com/guidovezzoni/gradle/smartproperties/extensions/BaseFlavor.kt
