+++ src/main/kotlin/com/guidovezzoni/gradle/smartproperties/extensions/Strings.kt
